--- a/opengever/repository/tests/test_reference_prefix_manager.py
+++ b/opengever/repository/tests/test_reference_prefix_manager.py
@@ -11,14 +11,13 @@ class TestReferencePrefixManager(IntegrationTestCase):
 
     def setUp(self):
         super(TestReferencePrefixManager, self).setUp()
-        self.login(self.administrator)
         # move repo1 to prefix 3 which leaves prefix 1 unused
         manager = ReferenceNumberPrefixAdpater(self.branch_repository)
         manager.set_number(self.leaf_repository, 3)
 
     @browsing
     def test_list_and_unlock_unused_prefixes(self, browser):
-        browser.login(self.administrator)
+        self.login(self.administrator, browser)
         browser.open(self.branch_repository, view='referenceprefix_manager')
         self.assertEquals(
             [['1', u'Vertr\xe4ge und Vereinbarungen', 'Unlock'],
@@ -39,8 +38,8 @@ class TestReferencePrefixManager(IntegrationTestCase):
 
     @browsing
     def test_manager_handles_deleted_repositories_correctly(self, browser):
+        self.login(self.administrator, browser)
         api.content.delete(obj=self.leaf_repository)
-        browser.login(self.administrator)
         browser.open(self.branch_repository, view='referenceprefix_manager')
         self.assertEquals(
             [['1', '-- Already removed object --', 'Unlock'],
@@ -49,7 +48,7 @@ class TestReferencePrefixManager(IntegrationTestCase):
 
     @browsing
     def test_manager_shows_default_message_when_no_repository_available(self, browser):
-        browser.login(self.administrator)
+        self.login(self.administrator, browser)
         browser.open(self.leaf_repository, view='referenceprefix_manager')
 
         self.assertEquals(
@@ -58,14 +57,13 @@ class TestReferencePrefixManager(IntegrationTestCase):
 
     @browsing
     def test_manager_is_hidden_from_users_without_permission(self, browser):
-        self.login(self.regular_user)
-        browser.login(self.regular_user)
+        self.login(self.regular_user, browser)
         with browser.expect_unauthorized():
             browser.open(self.branch_repository, view='referenceprefix_manager')
 
     @browsing
     def test_unlock_actions_are_journalized(self, browser):
-        browser.login(self.administrator)
+        self.login(self.administrator, browser)
         browser.open(self.branch_repository, view='referenceprefix_manager')
         browser.click_on('Unlock')
         statusmessages.assert_no_error_messages()
@@ -78,7 +76,7 @@ class TestReferencePrefixManager(IntegrationTestCase):
 
     @browsing
     def test_error_while_unlock_shows_statusmessage(self, browser):
-        browser.login(self.administrator)
+        self.login(self.administrator, browser)
         browser.open(self.branch_repository,
                      view='referenceprefix_manager?prefix=3')
         statusmessages.assert_message(