I've rewritten all<sup>[1]</sup> `opengever.repository` tests with the new fixture based integration testing and eliminated the use of the factory based functional testing in this package.

I've chosen `opengever.repository` because I could move forward quite fast because of it's simplicity, so that we have an example we can discuss about. Converting parts with more dependencies and complexity, such as SPV, will take more time.

### My Comments:
- The slow part when building objects is indexing those in the catalog.
- There are usually tests such as "can be created TTW". Since those are important and they create new objects, those are still slow(er).
- There is a print statement in layer setup, telling us how much time the fixture setup took.
- I think factories, when well designed, can help us a lot to test better. It is a more real enviroment, for example regardign security. We should not test with `Manager`-role but should use users from the fixture. This forces us to think about those things while writing a test.
- I did not only make tests compatible but rewrote those to be "better" from my subjective point of view. I removed tests which I believe were duplicates or bad because of other reasons.

------

### NEW Testing log (integration / fixtures)
```
ftw.testing PATCHING: Products.CMFQuickInstallerTool.events.handleProfileImportedEvent with noop
ftw.testing PATCHING: Products.CMFQuickInstallerTool.events.handleBeforeProfileImportEvent with noop
Running opengever.core.testing.opengever.core:integration tests:
  Set up plone.testing.zca.LayerCleanup in 0.001 seconds.
  Set up plone.testing.z2.Startup in 0.208 seconds.
  Set up plone.app.testing.layers.PloneFixture in 6.037 seconds.
  Set up ftw.testing.layer.ComponentRegistryIsolationLayer in 0.000 seconds.
  Set up opengever.core.testing.ContentFixtureLayer (fixture setup in 4.567s) in 15.070 seconds.
  Set up ftw.testbrowser.drivers.layers.DefaultDriverFixture:traversal library in 0.000 seconds.
  Set up opengever.core.testing.opengever.core:integration in 0.000 seconds.
  Running:
    6/48 (12.5%)error parsing dom, failure to add csrf token to response for url http://nohost/plone/ordnungssystem/rechnungsprufungskommission/delete_repository
error parsing dom, failure to add csrf token to response for url http://nohost/plone/ordnungssystem/rechnungsprufungskommission/delete_repository
error parsing dom, failure to add csrf token to response for url http://nohost/plone/ordnungssystem/rechnungsprufungskommission/delete_repository
error parsing dom, failure to add csrf token to response for url http://nohost/plone/ordnungssystem/rechnungsprufungskommission/delete_repository

  Ran 48 tests with 0 failures, 0 errors, 0 skipped in 16.178 seconds.
Running zope.testrunner.layer.UnitTests tests:
  Tear down opengever.core.testing.opengever.core:integration in 0.000 seconds.
  Tear down opengever.core.testing.ContentFixtureLayer in 0.008 seconds.
  Tear down ftw.testing.layer.ComponentRegistryIsolationLayer in 0.000 seconds.
  Tear down plone.app.testing.layers.PloneFixture in 0.068 seconds.
  Tear down plone.testing.z2.Startup in 0.004 seconds.
  Tear down plone.testing.zca.LayerCleanup in 0.002 seconds.
  Tear down ftw.testbrowser.drivers.layers.DefaultDriverFixture:traversal library in 0.000 seconds.
  Set up zope.testrunner.layer.UnitTests in 0.000 seconds.
  Running:

  Ran 3 tests with 0 failures, 0 errors, 0 skipped in 0.506 seconds.
Tearing down left over layers:
  Tear down zope.testrunner.layer.UnitTests in 0.000 seconds.
Total: 51 tests, 0 failures, 0 errors, 0 skipped in 39.030 seconds.
```

### OLD Testing log (functional / builder)
```
ftw.testing PATCHING: Products.CMFQuickInstallerTool.events.handleProfileImportedEvent with noop
ftw.testing PATCHING: Products.CMFQuickInstallerTool.events.handleBeforeProfileImportEvent with noop
Running opengever.core.testing.opengever.core:functional tests:
  Set up plone.testing.zca.LayerCleanup in 0.000 seconds.
  Set up plone.testing.z2.Startup in 0.172 seconds.
  Set up plone.app.testing.layers.PloneFixture in 5.159 seconds.
  Set up ftw.testing.layer.ComponentRegistryIsolationLayer in 0.000 seconds.
  Set up ftw.builder.testing.BuilderLayer in 0.000 seconds.
  Set up opengever.core.sqlite_testing.SQLiteMemoryFixture in 0.002 seconds.
  Set up opengever.core.testing.OpengeverFixture:SQLiteMemoryFixture in 11.159 seconds.
  Set up ftw.builder.testing.set_builder_session_factory in 0.000 seconds.
  Set up opengever.core.testing.opengever.core:functional in 0.000 seconds.
  Running:
    9/52 (17.3%)error parsing dom, failure to add csrf token to response for url http://nohost/plone/ordnungssystem/opengever-repository-repositoryfolder/delete_repository
error parsing dom, failure to add csrf token to response for url http://nohost/plone/ordnungssystem/opengever-repository-repositoryfolder/delete_repository
error parsing dom, failure to add csrf token to response for url http://nohost/plone/ordnungssystem/opengever-repository-repositoryfolder/delete_repository
error parsing dom, failure to add csrf token to response for url http://nohost/plone/ordnungssystem/opengever-repository-repositoryfolder/delete_repository

  Ran 52 tests with 0 failures, 0 errors, 0 skipped in 48.117 seconds.
Running opengever.core.testing.DossierTemplateLayer tests:
  Set up opengever.core.testing.DossierTemplateLayer in 0.033 seconds.
  Running:

  Ran 3 tests with 0 failures, 0 errors, 0 skipped in 2.589 seconds.
Running zope.testrunner.layer.UnitTests tests:
  Tear down opengever.core.testing.DossierTemplateLayer in 0.002 seconds.
  Tear down opengever.core.testing.opengever.core:functional in 0.000 seconds.
  Tear down opengever.core.testing.OpengeverFixture:SQLiteMemoryFixture in 0.003 seconds.
  Tear down opengever.core.sqlite_testing.SQLiteMemoryFixture in 0.000 seconds.
  Tear down ftw.testing.layer.ComponentRegistryIsolationLayer in 0.000 seconds.
  Tear down plone.app.testing.layers.PloneFixture in 0.071 seconds.
  Tear down plone.testing.z2.Startup in 0.005 seconds.
  Tear down plone.testing.zca.LayerCleanup in 0.002 seconds.
  Tear down ftw.builder.testing.set_builder_session_factory in 0.000 seconds.
  Tear down ftw.builder.testing.BuilderLayer in 0.000 seconds.
  Set up zope.testrunner.layer.UnitTests in 0.000 seconds.
  Running:

  Ran 3 tests with 0 failures, 0 errors, 0 skipped in 0.579 seconds.
Tearing down left over layers:
  Tear down zope.testrunner.layer.UnitTests in 0.000 seconds.
Total: 58 tests, 0 failures, 0 errors, 0 skipped in 1 minutes 9.431 seconds.
```

---

@4teamwork/gever Please take a look if you want 😉 
Do we need to optimize `bin/mtest` because of the new layer at some point?

---

<sup>[1]</sup> I didn't change some mock-based unit tests as they seem to do the job.