@jone 2017-06-27
I have updated the `self.login()` to also support logging in the browser. I have updated the tests accordingly.
I think it is ready for merge now.